--- Skillverse/backend/.gitignore
@@ -1,11 +1,0 @@
-node_modules
-.env
-coverage
-coverage.json
-typechain
-typechain-types
-
-# Hardhat files
-cache
-artifacts
-
