--- Skillverse/backend/hardhat.config.js
@@ -1,6 +1,0 @@
-require("@nomicfoundation/hardhat-toolbox");
-
-/** @type import('hardhat/config').HardhatUserConfig */
-module.exports = {
-  solidity: "0.8.18",
-};
