--- Skillverse/backend/package.json
@@ -1,8 +1,0 @@
-{
-  "devDependencies": {
-    "@nomicfoundation/hardhat-toolbox": "^2.0.2",
-    "hardhat": "^2.14.0",
-    "prettier": "^2.8.8",
-    "prettier-plugin-solidity": "^1.1.3"
-  }
-}
