--- Skillverse/frontend/src/index.css
@@ -1,83 +1,0 @@
-/* index.css */
-* {
-    margin: 0;
-    box-sizing: border-box;
-  }
-  
-  body {
-    background-color: #0f1546;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-  }
-  
-  /* LOGIN SECTION */
-  
-  .login-section {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .login-btn {
-    background: #97e3fe;
-    color: #fff;
-    padding: 13px 35px;
-    font-size: 24px;
-    border: none;
-    font-weight: 600;
-    cursor: pointer;
-  }
-  
-  .instruction {
-    text-align: center;
-    color: #21bf96;
-    color: #feba35;
-    margin: 1rem 0;
-  }
-  
-  /* DASHBOARD SECTION */
-  
-  .dashboard-section {
-    display: none;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-  
-  .wallet-status {
-    font-size: 54px;
-    color: #21bf96;
-    letter-spacing: 1.5px;
-  }
-  
-  .wallet-address-heading,
-  .wallet-balance-heading {
-    color: white;
-    letter-spacing: 1.5px;
-    margin-top: 1rem;
-    text-align: center;
-  }
-  
-  .wallet-balance,
-  .wallet-address {
-    color: #feba35;
-    letter-spacing: normal;
-    display: block;
-    margin-top: 1rem;
-    background: #000;
-    padding: 8px;
-    border-radius: 19px;
-  }
-  
-  .logout-btn {
-    color: white;
-    background: #cc0000;
-    padding: 13px 35px;
-    font-size: 24px;
-    border: none;
-    font-weight: 600;
-    cursor: pointer;
-    margin-top: 40px;
-  }
-
