--- Skillverse/project/contacts/src/components/Welcome.css
@@ -1,22 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Orbitron:wght@600&family=Raleway:ital,wght@0,300;1,300&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Orbitron:wght@600&family=Raleway:ital,wght@1,300&display=swap');
-.Title {
-    color: #f394f8;
-    font-size: 64px;
-    font-family: 'Orbitron', sans-serif;
-}
-
-.Introduction {
-    align-items: center;
-    justify-content: center;
-    color: #818de0;
-    font-size: 24px;
-    font-family: 'Raleway', sans-serif;
-}
-
-.Welcome_Section {
-    align-items: center;
-    justify-content: center;
-    display: flex;
-    flex-direction: column;
-}
